requirements.txt
# LangChain 1.x agent stack (pinned for reproducible installs)
langchain>=1.0.0,<2.0.0
langchain-openai>=0.3.0
langchain-groq>=0.2.0
langchain-community>=0.3.0
langchain-core>=1.0.0
langgraph>=1.0.0
langgraph-checkpoint>=2.0.0
langgraph-prebuilt>=1.0.0
langchain-experimental>=0.3.0
langchain-classic>=1.0.0

# RAG / vector store
faiss-cpu>=1.8.0
langchain-text-splitters>=0.3.0
sentence-transformers>=3.0.0

# Search & APIs
duckduckgo-search>=6.0.0
ddgs>=9.0.0
tavily-python>=0.5.0

# App
python-dotenv>=1.0.0
streamlit>=1.33.0
pypdf>=4.0.0
